Added EDA and imputed data for Age variable using MLR
Use train_with_age moving forward! This dataset contains imputed values of age. I also changed the Pclass, SibSp, and Parch variables to categorical (by converting them to strings), but this might not be the case in other regressions! May be worth reverting back (was just useful when I was imputing age)